requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pywebview>=5.3
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Image Processing
opencv-python-headless>=4.9.0
Pillow>=10.0
numpy>=1.26.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.6.0
pytesseract>=0.3.10

# Scanner
comtypes>=1.4.0

# Printer (Windows GDI)
pywin32>=306

# NFC (PC/SC smart card access)
pyscard>=2.0.7

# Card API
httpx>=0.27.0

# Engraving
svgwrite>=1.4.3
lxml>=5.0
qrcode[pil]>=7.4

# Security / Crypto
cryptography>=42.0
bcrypt>=4.0

# Cloud Storage
boto3>=1.34.0

# Cloud Database
psycopg2-binary>=2.9.9

# Reports / Templates
jinja2>=3.1.0

# Utilities
python-multipart>=0.0.9
Levenshtein>=0.25.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
